p5

+++ Character/Player/Goblin.cc
@@ -19,10 +19,10 @@ void Goblin::attackIt(std::shared_ptr<Halfling> e){
         double damage = ceil((100/(100+d)) * getAttack());
         e->updateDamage(damage);
         update_message("PC deals ");
-        update_message(to_string(damage));
+        update_message(to_string(static_cast<int>(damage)));
         update_message(" damage to Halfling");
         update_message("(");
-        update_message(std::to_string(e->getHP()));
+        update_message(std::to_string(static_cast<int>(e->getHP())));
         update_message("/");
         update_message(std::to_string(e->getMaxHP()));
         update_message(")");
@@ -35,10 +35,10 @@ void Goblin::attackIt(std::shared_ptr<Merchant> e){
     double damage = ceil((100/(100+d)) * getAttack());
     e->updateDamage(damage);
     update_message("PC deals ");
-    update_message(to_string(damage));
+    update_message(to_string(static_cast<int>(damage)));
     update_message(" damage to Merchant");
     update_message("(");
-    update_message(std::to_string(e->getHP()));
+    update_message(std::to_string(static_cast<int>(e->getHP())));
     update_message("/");
     update_message(std::to_string(e->getMaxHP()));
     update_message(")");
